requirements.txt
Django
celery
dj-database-url
django-configurations
django-filter
django-celery-results
django-redis
djangorestframework
markdown
mohawk
pgpy
psycopg2
redis
taskcluster
